Support for YouTube videos.
When a user provides a link to a YouTube video, download subtitles and
summarize them the same way we do an article.

Although Google provides YouTube APIs it adds complexity to auth calls
to YouTube. It requires to build a consent screen for the app, show it
to a user and only then get a key. It can be done once and the key saved
for later, but GC is glitching and I wans't able to pass it through.

Instead the bot will crawl the page and extract a link to subtitles from
it. It doesn't require credentials for public content. We might switch
to YT API later.