build: Make target to run processes in parallel
There is a tool called `parallel` that suppose to handle multiple
processes run in parallel. It has a bunch of neat features, but
debugging an issue was too hard.

Good old make can run the same processe in parallel, easy to use and
set up. However, it can't run same multiple target cocurrently, so
there is a workaround to make as many targets as processe needed.

WARNING! User tabs in make file, spaces messes things up!