fix: Do not reply to voice messages automatically
This is a temporary fix to avoid transcribing voice messages when
someone replies to a one. Due to a bug the bot automatically transcribes
all voice messages in the group... and it looks like people like it.

This is until the bot has settings page to allows to toggle auto
transcribe